+++ app/ante/ante.go
@@ -168,12 +168,12 @@ func IsSystemTx(tx sdk.Tx, isAuthorizedSigner func(string) bool) bool {
 		}
 	}
 	switch innerMsg.(type) {
-	case *cctxtypes.MsgGasPriceVoter,
+	case *cctxtypes.MsgVoteGasPrice,
 		*cctxtypes.MsgVoteOnObservedInboundTx,
 		*cctxtypes.MsgVoteOnObservedOutboundTx,
 		*cctxtypes.MsgAddToOutTxTracker,
+		*observertypes.MsgVoteBlockHeader,
 		*observertypes.MsgVoteTSS,
-		*observertypes.MsgAddBlockHeader,
 		*observertypes.MsgAddBlameVote:
 		signers := innerMsg.GetSigners()
 		if len(signers) == 1 {