+++ app/ante/ante_test.go
@@ -32,12 +32,12 @@ func (mah *MockAnteHandler) AnteHandle(ctx sdk.Context, _ sdk.Tx, _ bool) (sdk.C
 
 func TestIsSystemTx(t *testing.T) {
 	// system tx types:
-	//      *cctxtypes.MsgGasPriceVoter,
+	//      *cctxtypes.MsgVoteGasPrice,
 	//		*cctxtypes.MsgVoteOnObservedInboundTx,
 	//		*cctxtypes.MsgVoteOnObservedOutboundTx,
 	//		*cctxtypes.MsgAddToOutTxTracker,
+	//		*observertypes.MsgVoteBlockHeader,
 	//		*observertypes.MsgVoteTSS,
-	//		*observertypes.MsgAddBlockHeader,
 	//		*observertypes.MsgAddBlameVote:
 	buildTxFromMsg := func(msg sdk.Msg) sdk.Tx {
 		txBuilder := app.MakeEncodingConfig().TxConfig.NewTxBuilder()
@@ -187,8 +187,8 @@ func TestIsSystemTx(t *testing.T) {
 			true,
 		},
 		{
-			"MsgAddBlockHeader",
-			buildTxFromMsg(&observertypes.MsgAddBlockHeader{
+			"MsgVoteBlockHeader",
+			buildTxFromMsg(&observertypes.MsgVoteBlockHeader{
 				Creator: sample.AccAddress(),
 			}),
 			isAuthorized,
@@ -196,8 +196,8 @@ func TestIsSystemTx(t *testing.T) {
 			true,
 		},
 		{
-			"MsgExec{MsgAddBlockHeader}",
-			buildAuthzTxFromMsg(&observertypes.MsgAddBlockHeader{
+			"MsgExec{MsgVoteBlockHeader}",
+			buildAuthzTxFromMsg(&observertypes.MsgVoteBlockHeader{
 				Creator: sample.AccAddress(),
 			}),
 			isAuthorized,