+++ e2e/runner/setup_zeta.go
@@ -214,6 +214,13 @@ func (runner *E2ERunner) SetupBTCZRC20() {
 	runner.BTCZRC20 = BTCZRC20
 }
 
+// EnableVerificationFlags enables the verification flags on ZetaChain
+func (runner *E2ERunner) EnableVerificationFlags() error {
+	runner.Logger.Print("⚙️ enabling verification flags for block headers")
+
+	return runner.ZetaTxServer.EnableVerificationFlags(e2eutils.FungibleAdminName)
+}
+
 // FundEmissionsPool funds the emissions pool on ZetaChain with the same value as used originally on mainnet (20M ZETA)
 func (runner *E2ERunner) FundEmissionsPool() error {
 	runner.Logger.Print("⚙️ funding the emissions pool on ZetaChain with 20M ZETA (%s)", txserver.EmissionsPoolAddress)