+++ x/crosschain/genesis.go
@@ -50,6 +50,8 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 	for _, elem := range genState.FinalizedInbounds {
 		k.SetFinalizedInbound(ctx, elem)
 	}
+
+	k.SetRateLimiterFlags(ctx, genState.RateLimiterFlags)
 }
 
 // ExportGenesis returns the crosschain module's exported genesis.
@@ -86,6 +88,11 @@ func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 		genesis.ZetaAccounting = amount
 	}
 	genesis.FinalizedInbounds = k.GetAllFinalizedInbound(ctx)
+
+	rateLimiterFlags, found := k.GetRateLimiterFlags(ctx)
+	if found {
+		genesis.RateLimiterFlags = rateLimiterFlags
+	}
 
 	return &genesis
 }