+++ x/crosschain/types/authz_test.go
@@ -8,12 +8,12 @@ import (
 )
 
 func TestGetAllAuthzZetaclientTxTypes(t *testing.T) {
-	require.Equal(t, []string{"/zetachain.zetacore.crosschain.MsgGasPriceVoter",
+	require.Equal(t, []string{"/zetachain.zetacore.crosschain.MsgVoteGasPrice",
 		"/zetachain.zetacore.crosschain.MsgVoteOnObservedInboundTx",
 		"/zetachain.zetacore.crosschain.MsgVoteOnObservedOutboundTx",
 		"/zetachain.zetacore.crosschain.MsgAddToOutTxTracker",
 		"/zetachain.zetacore.observer.MsgVoteTSS",
 		"/zetachain.zetacore.observer.MsgAddBlameVote",
-		"/zetachain.zetacore.observer.MsgAddBlockHeader"},
+		"/zetachain.zetacore.observer.MsgVoteBlockHeader"},
 		crosschaintypes.GetAllAuthzZetaclientTxTypes())
 }