+++ x/lightclient/client/cli/query_verification_flags.go
@@ -1,0 +1,37 @@
+package cli
+
+import (
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/spf13/cobra"
+	"github.com/zeta-chain/zetacore/x/lightclient/types"
+)
+
+func CmdShowVerificationFlags() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "show-verification-flags",
+		Short: "Show the verification flags",
+		Args:  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			params := &types.QueryVerificationFlagsRequest{}
+
+			res, err := queryClient.VerificationFlags(cmd.Context(), params)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}