+++ x/lightclient/types/verification_flags.go
@@ -1,0 +1,10 @@
+package types
+
+// DefaultVerificationFlags returns the default verification flags.
+// By default, everything disabled.
+func DefaultVerificationFlags() VerificationFlags {
+	return VerificationFlags{
+		EthTypeChainEnabled: false,
+		BtcTypeChainEnabled: false,
+	}
+}