+++ zetaclient/zetabridge/zetacore_bridge.go
@@ -197,7 +197,7 @@ func (b *ZetaCoreBridge) GetKeys() *keys.Keys {
 
 // UpdateZetaCoreContext updates core context
 // zetacore stores core context for all clients
-func (b *ZetaCoreBridge) UpdateZetaCoreContext(coreContext *corecontext.ZetaCoreContext, init bool) error {
+func (b *ZetaCoreBridge) UpdateZetaCoreContext(coreContext *corecontext.ZetaCoreContext, init bool, sampledLogger zerolog.Logger) error {
 	bn, err := b.GetZetaBlockHeight()
 	if err != nil {
 		return err
@@ -222,7 +222,6 @@ func (b *ZetaCoreBridge) UpdateZetaCoreContext(coreContext *corecontext.ZetaCore
 	var newBTCParams *observertypes.ChainParams
 
 	// check and update chain params for each chain
-	sampledLogger := b.logger.Sample(&zerolog.BasicSampler{N: 10})
 	for _, chainParam := range chainParams {
 		if !chainParam.GetIsSupported() {
 			sampledLogger.Info().Msgf("Chain %d is not supported yet", chainParam.ChainId)
@@ -267,7 +266,24 @@ func (b *ZetaCoreBridge) UpdateZetaCoreContext(coreContext *corecontext.ZetaCore
 		return err
 	}
 
-	coreContext.Update(keyGen, newChains, newEVMParams, newBTCParams, tssPubKey, crosschainFlags, init, b.logger)
+	verificationFlags, err := b.GetVerificationFlags()
+	if err != nil {
+		b.logger.Info().Msg("Unable to fetch verification flags from zetabridge")
+		return err
+	}
+
+	coreContext.Update(
+		keyGen,
+		newChains,
+		newEVMParams,
+		newBTCParams,
+		tssPubKey,
+		crosschainFlags,
+		verificationFlags,
+		init,
+		b.logger,
+	)
+
 	return nil
 }